+++ filament/include/filament/View.h
@@ -62,6 +62,13 @@ class UTILS_PUBLIC View : public FilamentAPI {
 public:
     using TargetBufferFlags = backend::TargetBufferFlags;
 
+    enum class QualityLevel : int8_t {
+        LOW,
+        MEDIUM,
+        HIGH,
+        ULTRA
+    };
+
     /**
      * Dynamic resolution can be used to either reach a desired target frame rate
      * by lowering the resolution of a View, or to increase the quality when the
@@ -115,11 +122,35 @@ public:
         bool homogeneousScaling = false;                //!< set to true to force homogeneous scaling
     };
 
-    enum class QualityLevel : int8_t {
-        LOW,
-        MEDIUM,
-        HIGH,
-        ULTRA
+    /**
+     * Options to control the bloom effect
+     *
+     * enabled:     Enable or disable the bloom post-processing effect. Disabled by default.
+     * levels:      Number of successive blurs to achieve the blur effect, the minimum is 3 and the
+     *              maximum is 12. This value together with resolution influences the spread of the
+     *              blur effect. This value can be silently reduced to accommodate the original
+     *              image size.
+     * resolution:  Resolution of bloom's minor axis. The minimum value is 2^levels and the
+     *              the maximum is lower of the original resolution and 4096. This parameter is
+     *              silently clamped to the minimum and maximum.
+     *              It is highly recommended that this value be smaller than the target resolution
+     *              after dynamic resolution is applied (horizontally and vertically).
+     * strength:    how much of the bloom is added to the original image. Between 0 and 1.
+     * blendMode:   Whether the bloom effect is purely additive (false) or mixed with the original
+     *              image (true).
+     * anamorphism: Bloom's aspect ratio (x/y), for artistic purposes.
+     */
+    struct BloomOptions {
+        enum class BlendMode : uint8_t {
+            ADD,           //!< Bloom is modulated by the strength parameter and added to the scene
+            INTERPOLATE    //!< Bloom is interpolated with the scene using the strength parameter
+        };
+        float strength = 0.10f;                  //!< Between 0.0 and 1.0
+        uint32_t resolution = 360;               //!< Resolution of minor axis (2^levels to 4096)
+        float anamorphism = 1.0f;                //!< Bloom x/y aspect-ratio (1/32 to 32)
+        uint8_t levels = 6;                      //!< number of blur levels (3 to 12)
+        BlendMode blendMode = BlendMode::ADD;    //!< How the bloom effect is applied
+        bool enabled = false;                    //!< enable or disable bloom
     };
 
     /**
@@ -451,6 +482,20 @@ public:
     ToneMapping getToneMapping() const noexcept;
 
     /**
+     * Enables or disables bloom in the post-processing stage. Disabled by default.
+     *
+     * @param bloom options
+     */
+    void setBloomOptions(BloomOptions options) noexcept;
+
+    /**
+     * Queries the bloom options.
+     *
+     * @return the current bloom options for this view.
+     */
+    BloomOptions getBloomOptions() const noexcept;
+
+    /**
      * Enables or disables dithering in the post-processing stage. Enabled by default.
      *
      * @param dithering dithering type
@@ -518,6 +563,7 @@ public:
      * Enables or disables post processing. Enabled by default.
      *
      * Post-processing includes:
+     *  - Bloom
      *  - Tone-mapping & gamma encoding
      *  - Dithering
      *  - MSAA
@@ -529,7 +575,7 @@ public:
      *
      * @param enabled true enables post processing, false disables it.
      *
-     * @see setToneMapping, setAntiAliasing, setDithering, setSampleCount
+     * @see setBloomOptions, setToneMapping, setAntiAliasing, setDithering, setSampleCount
      */
     void setPostProcessingEnabled(bool enabled) noexcept;